refactor into failures.go, improve documentation
